Add auth service and rename column
