+++ src/main/kotlin/com/ledahl/apps/recieppyapi/auth/context/ContextBuilder.kt
@@ -1,6 +1,6 @@
 package com.ledahl.apps.recieppyapi.auth.context
 
-import com.ledahl.apps.recieppyapi.service.UserService
+import com.ledahl.apps.recieppyapi.service.AuthService
 import graphql.servlet.DefaultGraphQLContextBuilder
 import graphql.servlet.GraphQLContext
 import org.springframework.beans.factory.annotation.Autowired
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Component
 import javax.servlet.http.HttpServletRequest
 
 @Component
-class ContextBuilder(@Autowired private val userService: UserService): DefaultGraphQLContextBuilder() {
+class ContextBuilder(@Autowired private val authService: AuthService): DefaultGraphQLContextBuilder() {
     override fun build(httpServletRequest: HttpServletRequest?): GraphQLContext {
         val authentication = SecurityContextHolder.getContext().authentication
-        val user = userService.handleUserAuthentication(authentication)
+        val user = authService.handleUserAuthentication(authentication)
 
         return AuthContext(request = httpServletRequest, user = user)
     }
