+++ src/main/kotlin/com/ledahl/apps/recieppyapi/model/User.kt
@@ -3,7 +3,7 @@ package com.ledahl.apps.recieppyapi.model
 import com.ledahl.apps.recieppyapi.model.enums.UserRole
 
 data class User(val id: Long = 0,
-                val externalId: String,
+                val subject: String,
                 val firstName: String,
                 val lastName: String,
                 val email: String? = null,
