+++ src/main/resources/db/migration/V16__Rename_external_id_column_on_user_account.sql
@@ -1,0 +1,2 @@
+ALTER TABLE user_account
+    RENAME COLUMN external_id TO subject;
